Features, Bug Fixes and Upgrade
upgrade to Flutter 3.44.8 / Dart 3.12.2
Embedded Local SQL Query Engine for CSV & JSON (Feature 4)
Help topics for QR sharing, backup, and search
Updated Feature Analysis & Roadmap from 18-App Ecosystem Audit
Add green ticks and status icons to Section 5 feature matrix
fix all `prefer_initializing_formals` analyzer warnings
Guidelines conformance
run `dart format` across our own Dart code
The three remaining guidelines items
Android backup turned off
Feature 7 (AirQR) scope, payload caps, and non-goals written into the roadmap
Optical air-gap transfer (AirQR) — Steps 1 and 2 built
Feature 7 (AirQR) marked as delivered in the roadmap
Workspace-Wide SQLite FTS5 Full-Text Search Indexer
Ephemeral / Self-Destructing Document Workspace (Feature 9)
Tamper-Evident Workspace Audit Log (Feature 8)
Implement Feature 12 — Zero-Knowledge Encrypted Backup Archive (.txdata)
Per-Document Biometric Vault (.txvault) & P2P Direct Document File Transfer
Multi-Cursor & Column Block Selection (Shared Editor Core)
Offline Privacy Shield & Intelligent PII Scrubbing Engine
Serverless P2P Live Document Diff & Delta Sync
Enhance Help Section with New Features
Fix Last Card Clipping in Help Section

+++ lib/core/editor/pinch_to_zoom_area.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../theme/theme_controller.dart';
-import '../theme/theme_settings.dart';
+import 'package:text_data/core/theme/theme_controller.dart';
+import 'package:text_data/core/theme/theme_settings.dart';
 
 /// A widget that wraps any file viewer/editor surface and intercepts pinch/scale
 /// gestures to dynamically change the app-wide font size scale (CLAUDE.md §3).
@@ -117,9 +117,10 @@ class _PinchToZoomAreaState extends ConsumerState<PinchToZoomArea> {
     if (baseScale == null || baseSpan == null || baseSpan < 1.0) {
       return; // avoid divide-by-near-zero
     }
-    final newScale = (baseScale * (_currentVerticalSpan() / baseSpan))
-        .clamp(ThemeSettings.minFontScale, ThemeSettings.maxFontScale)
-        .toDouble();
+    final newScale = (baseScale * (_currentVerticalSpan() / baseSpan)).clamp(
+      ThemeSettings.minFontScale,
+      ThemeSettings.maxFontScale,
+    );
     ref.read(themeControllerProvider.notifier).setFontScale(newScale);
   }
